added and reviewed some services

+++ src/main/java/examination/teacherAndStudents/dto/CartResponse.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class CartResponse {
     private Long profileId;
     private String itemName;
     private String itemPhoto;
-    private Double itemPrice;
+    private BigDecimal itemPrice;
     private Integer quantity;
     private String size;
     private boolean isCheckedOut;
